refactor guitar to guitar_class to help track learning milestones
