+++ prac_06/guitar_class.py
